third_party/WebKit/LayoutTests/fast/forms/radio/radio-group-keyboard-change-event-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=32013.

For manual testing, focus a radio button in the first group and use the arrow keys. Changing the checked radio button should fire change events.

a b c

For manual testing, focus a radio button in the second group and use the arrow keys. Change events should still be dispatched but the checked radio should not change.

d e f

Test for https://code.google.com/p/chromium/issues/detail?id=556677.

For manual testing, focus a radio button in the first group and use the arrow keys. Changing the radio button should fire change events in the direction of left to right.

x y z

b dispatched change event
c dispatched change event
PASS: a is not checked
PASS: b is not checked
PASS: c is checked
PASS: d is checked
PASS: e is not checked
PASS: f is not checked
z dispatched change event
y dispatched change event
PASS: x is not checked
PASS: y is checked
PASS: z is not checked
PASS: x is not checked
PASS: y is checked
PASS: z is not checked
